bot-service/requirements.txt
# bot-service/requirements.txt
aiogram
python-dotenv
aiohttp